+++ plugins/modules/ksconf_app_manifest.py
@@ -18,12 +18,13 @@ import json
 import os
 import time
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, Tuple
 
 from ansible.module_utils.basic import AnsibleModule
 
 from ansible_collections.cdillc.splunk.plugins.module_utils.ksconf_shared import (
-    __version__ as collection_version, check_ksconf_version)
+    SIDELOAD_STATE_FILE, __version__ as collection_version,
+    check_ksconf_version)
 
 
 try:
@@ -50,11 +51,11 @@ ksconf_warn_version_text = ".".join(f"{i}" for i in ksconf_warn_version)
 DOCUMENTATION = r'''
 ---
 module: ksconf_app_manifest
-short_description: Splunk app manifest
+short_description: Splunk app manifest (Private)
 version_added: '0.22.0'
 author: Lowell C. Alleman (@lowell80)
 description:
-  - This is used internally by the M(ksconf_app_sideload).
+  - This is used internally by the M(cdillc.splunk.ksconf_app_sideload).
     This is not currently intended to be used directly by users.
 
 requirements:
@@ -65,28 +66,25 @@ options:
     description: Path to Splunk application
     type: path
     required: true
-  manifest:
-    description: Location of manifest file
+  state_file:
+    description: Location of the state file
     type: path
-    required: true
+    required: false
+  rebuild_manifest:
+    description: Rebuild the manifest information when missing from or corrupted with the state file.
+    type: bool
+    default: false
+
+extends_documentation_fragment:
+    - action_common_attributes
 
 attributes:
-  action:
-    support: full
-  async:
-    support: none
-  bypass_host_loop:
-    support: none
   check_mode:
-    support: full
+    support: none
   diff_mode:
     support: none
   platform:
     platforms: posix
-  safe_file_operations:
-    support: none
-  vault:
-    support: none
 
 notes:
   - Requires ksconf on the target host.
@@ -95,10 +93,10 @@ notes:
 
 
 EXAMPLES = r'''
-- name: Extract ta-nix.tgz into /opt/splunk/etc/apps
+- name: Don't do this.  This module is not public
   cdillc.splunk.ksconf_app_manifest:
     app_dir: /opt/splunk/etc/apps/myapp
-    state_file: app_dir: /opt/splunk/etc/apps/myapp/.ksconf_sideload.json
+    state_file: /opt/splunk/etc/apps/myapp/.ksconf_sideload.json
 
 '''
 
@@ -106,25 +104,61 @@ EXAMPLES = r'''
 RETURN = r'''
 app_dir:
   description:  Expanded application directory
-    - The source archive's path.
   returned: always
   type: str
-  sample: "/home/paul/test.tar.gz"
+  sample: "/opt/splunk/etc/app/fire_brigade"
+
+state_file:
+  description: Relative path to the json state tracking file where installation state, source hash,
+              and application manifest is stored.
+  returned: always
+  type: str
+  sample: /opt/splunk/etc/app/fire_brigade/.ksconf_sideload.json
+
+state_init:
+  description:
+    - Health indicator of first attempt at reading from I(present).
+    - >
+        Expect values such as:
+    - C(present) - state present and includes manifest,
+    - C(old-version) - state present with no manifest,
+    - C(corrupted) - unable to decode json,
+    - C(missing) - no state file present, or
+    - C(error) unexpected error attempting to load I(state_file).
+  returned: always
+  type: str
+
+result:
+  description:
+    - >
+      Final status of the state file.  Values include:
+    - C(loaded) when state & manifest successfully loaded from I(state_file),
+    - C(created) when state file was created from scratch,
+    - C(rebuilt) when existing state file was updated with new manifest, or
+    - C(no-manifest) when the manifest could not be loaded and I(rebuild_manifest) is false.
+    - >
+      Any of the follows indicate a failure:
+    - C(no-app) when the I(app_dir) is missing,
+    - C(error) when an unexpected error occurred.
+  type: str
+  returned: always
+
+manifest:
+  description: >
+    Manifest objects.
+    See L(AppManifest,https://ksconf.readthedocs.io/en/latest/api/ksconf.app.html#ksconf.app.manifest.AppManifest)
+  returned: when manifest is present or built
+  type: dict
+
 state:
   description: State of the state file / manifest
-  returned: always
-  type: str
-  sample:
-# state_file:
-#  description: Relative path to the json state tracking file where installation state, source hash,
-#               and application manifest is stored.
-#  returned: always
-#  type: str
-#  sample: fire_brigade/.ksconf_sideload.json
+  returned: on success
+  type: dict
+
 '''
 
 
-def get_app_manifest(state_file: Path) -> AppManifest:
+def get_app_manifest(state_file: Path) -> Tuple[AppManifest, dict, str]:
     from ksconf.app.manifest import AppManifest
     try:
         with open(state_file) as fp:
@@ -142,24 +176,27 @@ def get_app_manifest(state_file: Path) -> AppManifest:
 
 
 def build_app_manifest(app_dir: Path, state_file: Path) -> AppManifest:
+    """ Build a new app manifest for an existing app directory. """
     from ksconf.app.manifest import AppManifest
 
-    # Is this overcomplicated?   Is the better option:   lambda p: p.name != SIDELOAD_STATE_FILE
-    if state_file.exists() and state_file.is_relative_to(app_dir):
-        rel_statefile = state_file.relative_to(app_dir)
-
-        def filter_state_file(path):
-            return path != rel_statefile
-    else:
-        # Paths not relative no existing state file, therefore no filter needed
-        filter_state_file = None
+    if state_file.exists():
+        try:
+            rel_statefile = state_file.relative_to(app_dir)
+            def filter_state_file(path): return path != rel_statefile
+        except ValueError:
+            # State file not stored within the app directory.  Nothing to filter out
+            # NOTE: Can't use is_relative_to(); added in Python 3.9, still supporting 3.8
+            filter_state_file = None
+        else:
+            # No existing state file; nothing to filter (other deployment mechanism, or alternate state_file location?)
+            filter_state_file = None
 
     try:
         # Ksconf v0.13.4 adds 'filter_file'
         manifest = AppManifest.from_filesystem(app_dir, calculate_hash=True,
                                                filter_file=filter_state_file)
     except TypeError:
-        # Older ksconf.  Simply remove (and then restore) the manifest file
+        # Older ksconf.  Simply move manifest up a directory and restore when done.
         try:
             if state_file.exists():
                 temp_state = app_dir.parent.joinpath(f".{app_dir.name}.ksconf-manifest-rebuild.tmp")
@@ -200,6 +237,7 @@ def main():
             app_dir=dict(type='path', required=True),
             state_file=dict(type="path", required=False),
             rebuild_manifest=dict(type=bool, default=False),
+            raise_exception=dict(type=bool, default=False),        # Undocumented.  For internal debugging
         ),
         supports_check_mode=False,
     )
@@ -216,8 +254,14 @@ def main():
         module.warn(f"ksconf version {ksconf_version} is older than {ksconf_warn_version_text}.")
 
     app_dir = Path(module.params['app_dir'])
-    state_file = Path(module.params["state_file"])
+    state_file = module.params["state_file"]
     rebuild_manifest = module.params['rebuild_manifest']
+    raise_exception = module.params["raise_exception"]
+
+    if state_file:
+        state_file = Path(state_file)
+    else:
+        state_file = app_dir / SIDELOAD_STATE_FILE
 
     if module.check_mode:
         module.exit_json(msg="Check mode unsupported....  Please finish the implementation!")
@@ -227,25 +271,57 @@ def main():
         "state_file": os.fspath(state_file),
     }
     if not app_dir.is_dir():
-        module.fail_json(msg=f"App directory {app_dir} does not exists", result="no-app")
+        module.fail_json(msg=f"App directory {app_dir} does not exists", result="no-app", **results)
     if not os.access(app_dir, os.R_OK):
-        module.fail_json(msg=f"App directory {app_dir} is not readable", result="no-app")
+        module.fail_json(msg=f"App directory {app_dir} is not readable", result="no-app", **results)
 
-    # Is the state file missing (only a problem if rebuild_manifest is not present)
-    manifest, state, results["result"] = get_app_manifest(state_file)
+    manifest = state = None
+    try:
+        # Is the state_file missing (only a problem if rebuild_manifest is not present)
+        manifest, state, results["state_init"] = get_app_manifest(state_file)
+    except Exception as e:
+        results["state_init"] = "error"
+        module.warn(f"Unexpected exception during get_app_manifest(): {type(e).__name__}: {e}")
+        if raise_exception:
+            raise e
 
-    if not manifest and rebuild_manifest:
-        manifest = build_app_manifest(app_dir, state_file)
-        state = write_app_state(state_file, manifest, state)
-        results["result"] = "rebuilt"
+    if manifest:
+        results["result"] = "loaded"
+    else:
+        if rebuild_manifest:
+            # Attempt to build a new manifest
 
-    if not manifest:
-        module.fail_json(msg=f"State file {state_file} does not exist",
-                         result="no-manifest")
-        return
+            prev_mod_version = state.get('ansible_module_version', '?')
+            try:
+                manifest = build_app_manifest(app_dir, state_file)
+                state = write_app_state(state_file, manifest, state)
+                results["changed"] = True
+            except Exception as e:
+                results["result"] = "error"
+                module.fail_json(msg=f"Failed to build state file {state_file} ({results['state_init']}) "
+                                     f"from app {app_dir} due to exception: {type(e).__name__}: {e}", **results)
+                if raise_exception:
+                    raise e
+                return
+
+            if results["state_init"] == "old-version":
+                module.warn(f"Built new manifest due to version upgrade.  "
+                            f" {prev_mod_version} -> {collection_version}")
+            else:
+                module.warn(f"Built new manifest.  Previous state file was {results['state_init']}")
+
+            if results["state_init"] == "missing":
+                results["result"] = "created"
+            else:
+                results["result"] = "rebuilt"
+        else:
+            # No manifest information present; and rebuild is prohibited
+            results["result"] = "no-manifest"
 
     results["state"] = state
-    results["manifest"] = manifest.to_dict()
+    if manifest:
+        results["manifest"] = manifest.to_dict()
+
     module.exit_json(**results)